+++ tests/Lexer/TokensListTest.php
@@ -30,17 +30,22 @@ class TokensListTest extends TestCase
             new Token(' ', Token::TYPE_WHITESPACE),
             new Token('*', Token::TYPE_OPERATOR),
             new Token(' ', Token::TYPE_WHITESPACE),
-            new Token('FROM', Token::TYPE_KEYWORD),
+            new Token('FROM', Token::TYPE_KEYWORD, Token::FLAG_KEYWORD_RESERVED),
             new Token(' ', Token::TYPE_WHITESPACE),
             new Token('`test`', Token::TYPE_SYMBOL),
             new Token(' ', Token::TYPE_WHITESPACE),
+            new Token('WHERE', Token::TYPE_KEYWORD, Token::FLAG_KEYWORD_RESERVED),
+            new Token(' ', Token::TYPE_WHITESPACE),
+            new Token('name', Token::TYPE_NONE),
+            new Token('=', Token::TYPE_OPERATOR),
+            new Token('fa', Token::TYPE_NONE),
         ];
     }
 
     public function testBuild(): void
     {
         $list = new TokensList($this->tokens);
-        $this->assertEquals('SELECT * FROM `test` ', TokensList::build($list));
+        $this->assertEquals('SELECT * FROM `test` WHERE name=fa', TokensList::build($list));
     }
 
     public function testAdd(): void
@@ -60,6 +65,10 @@ class TokensListTest extends TestCase
         $this->assertEquals($this->tokens[2], $list->getNext());
         $this->assertEquals($this->tokens[4], $list->getNext());
         $this->assertEquals($this->tokens[6], $list->getNext());
+        $this->assertEquals($this->tokens[8], $list->getNext());
+        $this->assertEquals($this->tokens[10], $list->getNext());
+        $this->assertEquals($this->tokens[11], $list->getNext());
+        $this->assertEquals($this->tokens[12], $list->getNext());
         $this->assertNull($list->getNext());
     }
 
@@ -78,7 +87,22 @@ class TokensListTest extends TestCase
         $list = new TokensList($this->tokens);
         $this->assertEquals($this->tokens[0], $list->getNextOfType(Token::TYPE_KEYWORD));
         $this->assertEquals($this->tokens[4], $list->getNextOfType(Token::TYPE_KEYWORD));
+        $this->assertEquals($this->tokens[8], $list->getNextOfType(Token::TYPE_KEYWORD));
         $this->assertNull($list->getNextOfType(Token::TYPE_KEYWORD));
+    }
+
+    public function testGetNextOfTypeAndFlag(): void
+    {
+        $list = new TokensList($this->tokens);
+        $this->assertEquals($this->tokens[4], $list->getNextOfTypeAndFlag(
+            Token::TYPE_KEYWORD,
+            Token::FLAG_KEYWORD_RESERVED
+        ));
+        $this->assertEquals($this->tokens[8], $list->getNextOfTypeAndFlag(
+            Token::TYPE_KEYWORD,
+            Token::FLAG_KEYWORD_RESERVED
+        ));
+        $this->assertNull($list->getNextOfTypeAndFlag(Token::TYPE_KEYWORD, Token::FLAG_KEYWORD_RESERVED));
     }
 
     public function testGetNextOfTypeAndValue(): void
@@ -107,7 +131,7 @@ class TokensListTest extends TestCase
 
         // offsetExists($offset)
         $this->assertArrayHasKey(2, $list);
-        $this->assertArrayNotHasKey(8, $list);
+        $this->assertArrayNotHasKey(13, $list);
 
         // offsetUnset($offset)
         unset($list[2]);